+++ packages/infinity-modules/production-log/src/views/ProductionMachineView.vue
@@ -1,29 +1,27 @@
 <template>
-  <div>
-    <template v-if="loading">
-      Loading...
-    </template>
+  <div style="height:100%">
     <dashboard-toolbar-extension/>
-    <production-log-production-view/>
+    <template v-if="loading">
+      <production-loading />
+    </template>
+    <production-log-production-view v-else />
   </div>
 </template>
-
 <script>
-import { mapMutations, mapActions } from 'vuex';
+import { mapMutations, mapActions, mapState } from 'vuex';
 import DashboardToolbarExtension from '../components/core/DashboardToolbarExtension.vue';
 import ProductionLogProductionView from '../components/ProductionLogProductionView.vue';
-
+import ProductionLoading from '../components/ProductionLoading.vue';
 
 export default {
   name: 'ProductionMachineView',
   components: {
     ProductionLogProductionView,
     DashboardToolbarExtension,
+    ProductionLoading,
   },
-  data() {
-    return {
-      loading: false,
-    };
+  computed: {
+    ...mapState('productionLog', ['loading']),
   },
   async created() {
     await this.fetchBusinessHours();
@@ -32,7 +30,6 @@ export default {
     if (machines) {
       this.setExtendedHeader(true);
     }
-    this.loading = false;
   },
   watch: {
     logView(val) {