+++ packages/infinity/src/components/insights/InsightDetails.vue
@@ -4,16 +4,16 @@
       <span>You asked:&nbsp;</span>
       <span v-text="query.name"></span>
     </v-card-text>
-    <v-progress-linear :indeterminate="true"></v-progress-linear>
+    <v-progress-linear indeterminate v-if="loading"></v-progress-linear>
     <v-card-text class="text-center">
       <v-row align="center" no-gutters style="height: 100%">
         <v-col cols="12">
-          <div class="title font-weight-regular">
-            Insight detail here.
-          </div>
-          <div class="subheading">
-            Check back soon!
-          </div>
+          <highcharts
+            v-if="insightDetails
+              && insightDetails.type
+              && insightDetails.type.toUpperCase().includes('CHART')"
+            :options="insightDetails.chartOptions"
+          ></highcharts>
         </v-col>
       </v-row>
     </v-card-text>
@@ -26,7 +26,7 @@ import { mapState } from 'vuex';
 export default {
   name: 'InsightDetails',
   computed: {
-    ...mapState('insight', ['query']),
+    ...mapState('insight', ['query', 'window', 'loading', 'insightDetails']),
   },
 };
 </script>