+++ packages/infinity/src/plugins/highcharts.js
@@ -3,6 +3,8 @@ import Highcharts from 'highcharts';
 import HighchartsVue from 'highcharts-vue';
 import HighchartsMore from 'highcharts/highcharts-more';
 import xRange from 'highcharts/modules/xrange';
+import pareto from 'highcharts/modules/pareto';
+
 
 Highcharts.theme = {
   chart: {
@@ -19,4 +21,5 @@ Highcharts.theme = {
 Highcharts.setOptions(Highcharts.theme);
 HighchartsMore(Highcharts);
 xRange(Highcharts);
+pareto(Highcharts);
 Vue.use(HighchartsVue);